refactor(repository): split MedTrackRepository into focused repositories
- Add AuthRepository (login, token, session)
- Add MedicamentoRepository (fetch user/medicamentos, save medicamentos_v2, schedule notifications, confirm medication)
- Add ScanRepository (online scan, offline queue, pending uploads)
- Add LoginData.kt for login/sync result
- Remove MedTrackRepository.kt
- Update ViewModels to receive only relevant repositories
- Update ScanUpload Worker to use ScanRepository via Hilt EntryPoint

Ref #16
